Ask frontier for total queue count
Refactor database query as own package
