+++ go.mod
@@ -1,9 +1,17 @@
 module github.com/nlnwa/veidemann-metrics
 
-go 1.13
+go 1.14
 
 require (
-	github.com/prometheus/client_golang v1.6.0
-	github.com/prometheus/common v0.9.1
+	github.com/golang/protobuf v1.4.3
+	github.com/nlnwa/veidemann-api-go v1.0.0-beta18
+	github.com/opentracing/opentracing-go v1.2.0 // indirect
+	github.com/prometheus/client_golang v1.8.0
+	github.com/prometheus/common v0.15.0
+	github.com/sirupsen/logrus v1.7.0 // indirect
+	golang.org/x/crypto v0.0.0-20201117144127-c1f2f97bffc9 // indirect
+	golang.org/x/net v0.0.0-20201110031124-69a78807bb2b // indirect
+	golang.org/x/sys v0.0.0-20201119102817-f84b799fce68 // indirect
+	google.golang.org/grpc v1.33.1
 	gopkg.in/rethinkdb/rethinkdb-go.v6 v6.2.1
 )
